requirements.txt
asgiref==3.2.10
dj-database-url==0.5.0
Django>=3.1.10
django-ckeditor>=5.9.0
django-crispy-forms==1.9.2
django-heroku>=0.3.1
django-js-asset==1.2.2
Pillow>=8.1.2
python-decouple==3.3
python-dotenv==0.14.0
pytz==2020.1
sqlparse==0.3.1
whitenoise>=5.2.0
martor
django-modeltranslation
